Refactor: #197: transaction  - 파일 관련 메서드 util 분리 및 applcation.yml 환경변수 추가
