+++ backend/tiggle-root/tiggle/src/main/resources/application.yml.example
@@ -70,6 +70,13 @@ part:
         - image/jpeg
         - image/png
         - image/gif
+    transaction:
+      path: upload/transaction/
+      max-size: 10485760  # 10MB
+      allowed-types:
+        - image/jpeg
+        - image/png
+        - image/gif
 
 app:
   client-redirect-uri: https://tiggle.duckdns.org/login/success
